backend/requirements.txt
fastapi
uvicorn
langchain==0.2.16
langchain-community==0.2.16
langchain-groq
langchain-core
langchain-text-splitters
faiss-cpu
sentence-transformers
pypdf
python-multipart
python-dotenv